# Conflicts:
#	wire-ios-data-model/Source/Model/User/TypeOfUser.swift
#	wire-ios-data-model/Source/Model/User/ZMUser.swift
#	wire-ios-data-model/WireDataModel.xcodeproj/project.pbxproj

+++ wire-ios-data-model/Source/Model/User/TypeOfUser.swift
@@ -17,7 +17,7 @@
 //
 
 @objc(ZMTypeOfUser)
-public enum TypeOfUser: Int, CaseIterable, Sendable {
+public enum TypeOfUser: Int16, CaseIterable, Sendable {
     case regular
     case app
     case bot
